refactor: move defaults into separate module
Move default mappings and autocommands into a separate module and add
comments and docstrings to document each of the defaults.
